update notebooks, feather_v2 -> feather
